requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-jose==3.3.0
passlib==1.7.4
pyotp==2.8.0
pydantic==1.10.7
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
pymongo==4.3.3
email-validator==2.0.0.post2
python-multipart==0.0.6
aiofiles==23.1.0
python-dotenv==1.0.0